+++ server/src/Devilution.Server/Gameplay/AuthoritativeObjectEffect.cs
@@ -1,0 +1,9 @@
+namespace Devilution.Server.Gameplay;
+
+public enum AuthoritativeObjectEffectKind
+{
+    None,
+    Heal,
+    Damage,
+    Experience,
+}